Added some catch for the queries All interactions are made in the main thread

+++ src/main/java/cl/dreamit/elevateit/DataModel/DAO/UploadableDAO.java
@@ -6,5 +6,5 @@ import cl.dreamit.elevateit.DataModel.Entities.FullAccess.UploadableEntity;
 
 public interface UploadableDAO<T extends UploadableEntity> {
     public String getTable();
-    public List<T> getNewerThan(long lastID);
+    public List<T> getNewerThan(int lastID);
 }
